+++ data/transaction/status/interface.go
@@ -1,0 +1,18 @@
+package status
+
+import (
+	"github.com/multiversx/mx-chain-core-go/data/block"
+	"github.com/multiversx/mx-chain-core-go/data/transaction"
+)
+
+// StatusComputerHandler computes a transaction status
+type StatusComputerHandler interface {
+	ComputeStatusWhenInStorageKnowingMiniblock(miniblockType block.Type, tx *transaction.ApiTransactionResult) (transaction.TxStatus, error)
+	ComputeStatusWhenInStorageNotKnowingMiniblock(destinationShard uint32, tx *transaction.ApiTransactionResult) (transaction.TxStatus, error)
+	SetStatusIfIsRewardReverted(
+		tx *transaction.ApiTransactionResult,
+		miniblockType block.Type,
+		headerNonce uint64,
+		headerHash []byte,
+	) (bool, error)
+}
